🚚 Rename: 검색 컴포넌트 파일 구조 분리 (#227)
